Refactor: Move data fetching functions to dataHandlers.js and improve error handling
- Moved getFolders, getNotes, getEvents, getHashtags, getMentions, getTeamspaces
  from requestHandlers.js to new dataHandlers.js to break circular dependencies
- Moved RequestResponse type to shared/types.js for better type sharing
- Improved error handling in FormBrowserView for form submission responses
- Updated imports across affected files
